Move InternalPlugin commands
