+++ Sync/Sync.csproj
@@ -96,6 +96,8 @@
     <Compile Include="MessageFilter\FilterManager.cs" />
     <Compile Include="MessageFilter\FilterPriorityAttribute.cs" />
     <Compile Include="MessageFilter\MessageDispatcher.cs" />
+    <Compile Include="Plugins\BuildInPlugin\Commands\CommonCommand.cs" />
+    <Compile Include="Plugins\BuildInPlugin\Commands\PluginCommand.cs" />
     <Compile Include="Plugins\BuildInPlugin\InternalPlugin.cs" />
     <Compile Include="Tools\Updater.cs" />
     <Compile Include="Plugins\Plugin.cs" />
